remove interface file
