+++ .pre-commit-config.yaml
@@ -9,6 +9,8 @@ repos:
         # line too long and line before binary operator (black is ok with these)
         types:
           - python
+        args:
+          - "--max-line-length=90"
       - id: trailing-whitespace
       - id: end-of-file-fixer
       - id: check-yaml
