+++ vetiver/__init__.py
@@ -14,9 +14,8 @@ from .attach_pkgs import *  # noqa
 from .meta import *  # noqa
 from .write_docker import write_docker  # noqa
 from .write_fastapi import write_app  # noqa
-from .handlers._interface import create_handler, InvalidModelError  # noqa
-from .handlers.base import VetiverHandler  # noqa
+from .handlers.base import VetiverHandler, create_handler, InvalidModelError  # noqa
 from .handlers.sklearn import SKLearnHandler  # noqa
 from .handlers.torch import TorchHandler  # noqa
-from .rsconnect import deploy_rsconnect # noqa
-from .monitor import compute_metrics, pin_metrics, plot_metrics, _rolling_df # noqa
+from .rsconnect import deploy_rsconnect  # noqa
+from .monitor import compute_metrics, pin_metrics, plot_metrics, _rolling_df  # noqa
